% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_PCAs.R
\name{calc_PCAs}
\alias{calc_PCAs}
\title{Run PCA on gene expression count data}
\usage{
\code{
calc_PCAs(counts, cpm=TRUE, log2_transform=FALSE, ...)}
}
\arguments{
\item{counts}{matrix or data frame of gene expression counts, or an object from which counts can be extracted. Should have genes in rows and samples in columns}

\item{cpm}{logical, whether to transform the counts to counts-per-million (using \code{edgeR::cpm}) before running prcomp.}

\item{log2_transform}{logical, whether to transform the counts (using \code{log2(counts+1)}) before running prcomp.}

\item{...}{(optional) additional arguments passed to prcomp, such as \code{center} and \code{scale}.}
}
\value{
a prcomp object, with additional list elements containing the percent variance and strings for labeling plot axes.
}
\description{
This is a wrapper function for prcomp, designed for running PCA on RNAseq data. It transposes the input data, and adds two elements to the prcomp object, containing the following for each PC: the percent variance, and a character string for labeling plot axes.
}
