% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_gene_names_species.R
\name{convert_gene_symbols_species}
\alias{convert_gene_symbols_species}
\title{Convert gene identifiers from mouse to human or human to mouse, using orthologs}
\usage{
convert_gene_symbols_species(
  genes,
  species_from,
  species_to,
  return_format = "orthologs_only"
)
}
\arguments{
\item{genes}{character vector containing gene identifiers to convert}

\item{species_from}{character, the species of the input gene identifiers. Must be "mouse" or "human"; partial matches are accepted.}

\item{species_to}{character, the species of the output gene identifiers. Must be "mouse" or "human"; partial matches are accepted.}

\item{return_format}{character, the format of the output gene identifiers. Must be "orthologs_only" or "orig_and_ortholog"; partial matches are accepted.}
}
\value{
a vector of gene symbols, of length similar to the input genes, minus any genes that do not have orthologs in `species_to`.
}
\description{
Convert a vector of symbols from one species to another. Current options
exist only for mouse to human and  identifiers from one standard to another.
Mouse identifiers will be MGI symbols. Human identifiers will be HGNC
symbols. Multiple genes may be returned for any input gene, based on the
possible one-to-many mapping of gene orthologs. This function is adapted
from https://www.biostars.org/p/9567892/, and uses the MGI list of
orthologs at https://www.informatics.jax.org/downloads/reports/HOM_MouseHumanSequence.rpt
}
