% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_HGNC.R
\name{get_HGNC}
\alias{get_HGNC}
\title{Convert gene identifiers to HGNC gene symbols}
\usage{
\code{get_HGNC(genes, from="ensgene", type=NULL, use_annotables=TRUE)}
}
\arguments{
\item{genes}{character vector containing gene identifiers to convert}

\item{from}{the gene identifier class to convert from. Must match a corresponding variable in \code{annotables} or \code{biomaRt}.}

\item{type}{(optional) character vector containing types of genes to retain. If provided, genes with biotype not matching elements of this vector return NA. If NULL, all genes are retained.}

\item{use_annotables}{boolean, whether to use the annotables package. If annotables is not installed, the function defaults to using \code{biomaRt}.}
}
\value{
a vector of HGNC gene symbols, of length equivalent to the length of genes.
}
\description{
This is a simple wrapper for the more flexible function \code{convert_gene_names}. I created it to
facilitate compatability with old code. It simply calls \code{convert_gene_names} with the hard-coded
specification to return HGNC symbols, and some variable name translation.
It converts a vector of gene identifiers to HGNC gene symbols. The type of the input identifiers is
specified using \code{from}. By default, this function uses the \code{annotables} package. If annotables
is not installed, or if the user specifies not to use it, \code{biomaRt} is used instead. Genes can be
filtered by type; any genes not matching the specified type(s) will return NA.
}
