% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_outlier.R
\name{is_outlier}
\alias{is_outlier}
\title{Determine outliers}
\usage{
\code{is_outlier(x)}
}
\arguments{
\item{x}{numeric vector containing values test for outliers}
}
\value{
a boolean vector, containing TRUE for outlier elements and FALSE for all other elements.
}
\description{
Determine outliers of a numeric vector, based on the first and third quartile minus/plus 1.5 times the
inter-quartile range.
}

