% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logXYratio.R
\name{logXYratio}
\alias{logXYratio}
\title{Calculate ratio of total counts of genes mapping to the X and Y chromosome}
\usage{
logXYratio(
  counts,
  lib_cols = 1:ncol(counts),
  gene_ID = "symbol",
  use_annotables = TRUE
)
}
\arguments{
\item{counts}{a matrix or data frame containing the gene expression counts. Should have samples in columns and genes in rows. Row names must contain gene names, in the class standard matching \code{gene_ID}.}

\item{lib_cols}{a numeric vector, the indices of columns containing count data. Defaults to all columns in \code{counts}; can be adjusted if non-count columns are included in the \code{counts} object.}

\item{gene_ID}{the gene identifier class of the gene names. Must match a corresponding variable in \code{annotables} or \code{biomaRt}.}

\item{use_annotables}{boolean, whether to use the annotables package. If annotables is not installed, defaults to using \code{biomaRt}.}
}
\value{
a vector of ratios, with one element for each sample.
}
\description{
This function calculates the number of reads that map to the X and Y
chromosomes, and returns the (natural) log of the ratio of total X reads
to total Y reads. It can be used to verify or infer sex using RNAseq data.
Large values are likely to be from female samples; small values from male samples.
}
\details{
\code{counts} should be normalized or raw, but not log-transformed.
 It assumes that each column in the counts object corresponds to a library.
 If the counts object contains additional columns, the columns containing
 libraries must be indicated in \code{lib_cols}.
}
