% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PC_loadings.R
\name{plot_PC_loadings}
\alias{plot_PC_loadings}
\title{Plot PC loadings}
\usage{
\code{plot_PC_loadings(rotation_matrix)}
}
\arguments{
\item{rotation_matrix}{a PCA rotation or loading matrix. Should have genes in rows and PCs in columns.}
}
\description{
Plot the loadings of each variable on each principal component. Each variable is plotted as a colored line.
In the future, I may expand this function to take prcomp and princomp objects, and to provide additional
handles for manipulating the plots (e.g. number of variables and/or number of PCs to plot), outputting
to pdf, etc.
}
