% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_read_counts.R
\name{plot_read_counts}
\alias{plot_read_counts}
\title{Plot read counts by library}
\usage{
plot_read_counts(
  metrics,
  file_prefix = NULL,
  plotdims = c(9, 6),
  threshold_line = 5,
  n_lowcount = 20,
  color_by_var = NULL,
  color_by_var_levels = NULL,
  color_var_lab = NULL,
  my_cols = c("red", "blue"),
  na_col = "grey50",
  id_col = "lib.id",
  total_reads_col = "fastq_total_reads"
)
}
\arguments{
\item{metrics}{matrix or data frame containing values of metrics. Should have metrics in columns and libraries in rows.}

\item{file_prefix}{a character string. If provided, the function outputs pdfs of the plots, named "{file_prefix}_{plot_name}.pdf". If not provided, the function prints to a plotting window.}

\item{plotdims}{a numeric vector, the size (in inches) of the plotting object. Either the size of the pdfs, or the size of the plotting windows.}

\item{threshold_line}{numeric, the values (in millions of reads) at which to plot a horizontal line.}

\item{n_lowcount}{numeric, the number of libraries to include in the plot of low-count libraries}

\item{color_by_var}{(optional) character string or integer identifying the column in \code{metrics} to color bars by. If not provided, bars are plotted in grey.}

\item{color_by_var_levels}{(optional) character vector defining the order of elements in the variable used for coloring bars; this order is used for the plot legend and to match the order of colors (if provided). If not provided, levels are taken from the factor levels (if color_by_var is a factor), or else are ordered by order of appearance in \code{metrics}.}

\item{color_var_lab}{(optional) string to be used as the title for the color legend.}

\item{my_cols}{(optional) vector of colors to use for plotting. If color_by_var is not numeric, should be a vector with one color for each level of \code{color_by_var}; if the number of values supplied is less than the numer of levels in color_by_var, additional values are interpolated using colorRampPalette. By default, uses a range from blue to red. Future updates will handle numeric vectors for \code{color_by_var}, as follows. \code{color_by_var} is numeric, should have two elements, providing the start and end points for a continuous color scale (generated by \code{scale_color_gradient}).}

\item{na_col}{color to use for NA values of \code{color_by_var}.}

\item{id_col}{numeric or character, the number or name of the column containing the library identifiers. Used to plot identifiers of low-count libraries. Defaults to "lib.id".}
}
\description{
Output plots of read counts by library, at coarse and fine scales. Libraries areordered by the total
number of reads. These plots can optionally be output to pdfs by specifing \code{file_prefix}. The first
plot shows all libraries, with a line at \code{threshold_line} million reads. The second plot zooms in
on the \code{n_lowcount} libraries with the lowest counts.
}
