% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_read_counts.R
\name{plot_read_counts}
\alias{plot_read_counts}
\title{Plot read counts by library}
\usage{
\code{
plot_read_counts(metrics,
                 file_prefix=NULL, plotdims=c(9,6),
                 threshold_line=5, n_lowcount=20,
                 id_col="lib.id"
                 )}
}
\arguments{
\item{metrics}{matrix or data frame containing values of metrics. Should have metrics in columns and libraries in rows.}

\item{file_prefix}{a character string. If provided, the function outputs pdfs of the plots, named "{file_prefix}_{plot_name}.pdf". If not provided, the function prints to a plotting window.}

\item{plotdims}{a numeric vector, the size (in inches) of the plotting object. Either the size of the pdfs, or the size of the plotting windows.}

\item{threshold_line}{numeric, the values (in millions of reads) at which to plot a horizontal line.}

\item{n_lowcount}{numeric, the number of libraries to include in the plot of low-count libraries}

\item{id_col}{numeric or character, the number or name of the column containing the library identifiers. Used to plot identifiers of low-count libraries.}
}
\description{
Output plots of read counts by library, at coarse and fine scales. Libraries areordered by the total
number of reads. These plots can optionally be output to pdfs by specifing \code{file_prefix}. The first
plot shows all libraries, with a line at \code{threshold_line} million reads. The second plot zooms in
on the \code{n_lowcount} libraries with the lowest counts.
}

